aa
aa

+++ Task_1/Task_1_15/src/com/company/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static void main(String[] args) {
         // 4. Найти сумму квадратов первых 200 чисел.
-        int k=0;
+        long k=0;
         for (int i=1; i<=200; i++ ){
             k+=i*i;
         }
